fix(Product): changing product id to Integer wrapper type.

+++ src/main/java/com/ifsp/edu/sanca_dinner/domain/model/Product.java
@@ -7,12 +7,12 @@ import java.math.BigDecimal;
 @Getter
 public class Product {
 
-    private Long id;
+    private Integer id;
     private String name;
     private BigDecimal price;
     private String description;
 
-    public Product(Long id, String name, BigDecimal price, String description) {
+    public Product(Integer id, String name, BigDecimal price, String description) {
         this.id = id;
         setName(name);
         setPrice(price);
